refactor(jwt): move token interceptor class
